--- polymorphia-frontend/app/(logged-in)/test/types.ts
@@ -1,11 +1,0 @@
-import { ReactNode } from "react";
-
-export interface SpeedDialProps {
-  items: SpeedDialItem[];
-}
-
-export interface SpeedDialItem {
-  label: string;
-  icon: string;
-  modal: (onClose: () => void) => ReactNode;
-}
