+++ polymorphia-frontend/components/button/index.css
@@ -7,7 +7,7 @@
 }
 
 .button-with-border-dark {
-    @apply text-primary-dark border-primary-dark hover:bg-primary-dark hover:text-neutral-300;
+    @apply text-primary-dark border-primary-dark hover:bg-primary-dark hover:text-secondary-gray;
 }
 
 .button-with-border-light {
