+++ polymorphia-frontend/components/image-badge/index.css
@@ -2,6 +2,6 @@
 @import "../../styles/globals.css";
 
 .image-badge {
-    @apply absolute bg-neutral-300/80 flex-col-centered bottom-0 right-0 aspect-square overflow-hidden  bg-clip-padding z-50;
+    @apply absolute bg-secondary-gray/80 flex-col-centered bottom-0 right-0 aspect-square overflow-hidden  bg-clip-padding z-50;
     @apply dark:bg-secondary-dark;
 }
