+++ polymorphia-frontend/components/navigation/Line.tsx
@@ -1,3 +1,3 @@
 export default function Line() {
-  return <div className="w-full border-[1.5px] border-neutral-300"></div>;
+  return <div className="w-full border-[1.5px] border-secondary-gray"></div>;
 }
