+++ polymorphia-frontend/components/speed-dial/index.css
@@ -1,0 +1,19 @@
+@import "tailwindcss";
+@import "../../styles/globals.css";
+
+.speed-dial {
+    @apply flex flex-col items-center gap-y-3;
+}
+
+.speed-dial-button-wrapper {
+    @apply relative w-13 h-13;
+}
+
+.speed-dial-button {
+    @apply p-0 w-full h-full text-2xl flex-centered rounded;
+}
+
+.speed-dial-label {
+    @apply absolute right-10/12 top-1/2 -translate-y-1/2 mr-1 px-3 py-1 bg-secondary-dark text-secondary-gray;
+    @apply text-[1.5rem] whitespace-nowrap rounded transition-all duration-250 ease-out;
+}
