+++ polymorphia-frontend/components/speed-dial/index.tsx
@@ -2,7 +2,8 @@
 
 import { ReactNode, useState } from "react";
 import ButtonWithBorder from "@/components/button/ButtonWithBorder";
-import { SpeedDialProps } from "@/app/(logged-in)/test/types";
+import { SpeedDialProps } from "@/components/speed-dial/types";
+import "./index.css";
 
 export default function SpeedDial({ items }: SpeedDialProps) {
   const [hoveredIndex, setHoveredIndex] = useState<number | null>(null);
@@ -10,19 +11,18 @@ export default function SpeedDial({ items }: SpeedDialProps) {
 
   return (
     <>
-      <div className="flex flex-col items-center gap-y-3">
+      <div className="speed-dial">
         {items.map((item, i) => (
           <div
             key={i}
-            className={`flex justify-center group`}
             onMouseEnter={() => setHoveredIndex(i)}
             onMouseLeave={() => setHoveredIndex(null)}
           >
-            <div className="relative w-13 h-13">
+            <div className="speed-dial-button-wrapper">
               <ButtonWithBorder
                 text=""
                 size="lg"
-                className="!p-0 w-full h-full !text-2xl !flex-centered !rounded"
+                className="speed-dial-button"
                 icon={item.icon}
                 onClick={() => {
                   if (item.modal) {
@@ -31,7 +31,7 @@ export default function SpeedDial({ items }: SpeedDialProps) {
                 }}
               />
               <div
-                className={`absolute right-10/12 top-1/2 -translate-y-1/2 mr-1 px-3 py-1 bg-secondary-dark text-neutral-300 text-[1.5rem] whitespace-nowrap rounded transition-all duration-250 ease-out pointer-events-none ${hoveredIndex === i ? "opacity-100 -translate-x-3" : "opacity-0 translate-x-0"}`}
+                className={`speed-dial-label ${hoveredIndex === i ? "opacity-100 -translate-x-3" : "opacity-0 translate-x-0"}`}
               >
                 {item.label}
               </div>
@@ -39,7 +39,6 @@ export default function SpeedDial({ items }: SpeedDialProps) {
           </div>
         ))}
       </div>
-
       {activeModal}
     </>
   );
