+++ polymorphia-frontend/components/speed-dial/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+export interface SpeedDialProps {
+  items: SpeedDialItem[];
+}
+
+export interface SpeedDialItem {
+  label: string;
+  icon: string;
+  modal: (onClose: () => void) => ReactNode;
+}
