+++ polymorphia-frontend/styles/globals.css
@@ -15,6 +15,7 @@
   --color-primary-dark: #262626;
   --color-secondary-dark: #212121;
   --color-primary-gray: #a1a1a1;
+  --color-secondary-gray: #d4d4d4;
 }
 
 @utility flex-centered {
